chore: rename namespace and service
- rename 'School' namespace to 'Schools' to avoid clashes with the model 'School'
- rename ''ECTService' to 'Teacher'
