--- app/controllers/school/register_ect_controller.rb
@@ -1,4 +1,0 @@
-class School::RegisterECTController < ApplicationController
-  def start
-  end
-end
