+++ app/controllers/schools/register_ect_controller.rb
@@ -1,0 +1,6 @@
+module Schools
+  class RegisterECTController < ApplicationController
+    def start
+    end
+  end
+end
