+++ app/views/schools/home/index.html.erb
@@ -1,6 +1,6 @@
 <% page_data(title: "Early career teachers (ECT)", error: false) %>
 
-<%= govuk_button_link_to("Add an ECT", school_register_ect_start_path) %>
+<%= govuk_button_link_to("Add an ECT", schools_register_ect_start_path) %>
 <hr class="govuk-section-break--m"/>
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-one-third">
