+++ app/views/schools/register_ect/start.html.erb
