+++ config/routes.rb
@@ -61,7 +61,7 @@ Rails.application.routes.draw do
     end
   end
 
-  namespace :school do
+  namespace :schools do
     get "/home/ects", to: "home#index", as: :ects_home
     get "/register-ect/start", to: "register_ect#start", as: :register_ect_start
   end
