+++ spec/services/schools/teacher_spec.rb
@@ -1,7 +1,7 @@
-describe School::ECTService do
+describe Schools::Teacher do
   let(:school) { FactoryBot.create(:school) }
   let(:ect) { FactoryBot.create(:teacher) }
-  let(:mentor) { FactoryBot.create(:teacher) }
+  let(:mentor) { FactoryBot.create(:teacher, corrected_name: nil) }
   let(:mentor_period) { FactoryBot.create(:mentor_at_school_period, school:, teacher: mentor, started_on: 2.years.ago, finished_on: nil) }
   let(:ect_period) { FactoryBot.create(:ect_at_school_period, school:, teacher: ect, started_on: 2.years.ago, finished_on: nil) }
 
@@ -9,7 +9,7 @@ describe School::ECTService do
 
   before do
     FactoryBot.create(:training_period, mentor_at_school_period: mentor_period, ect_at_school_period: nil, started_on: 1.year.ago, finished_on: nil)
-    FactoryBot.create(:mentorship_period, mentor: mentor_period, mentee: ect_period, started_on: 1.year.ago)
+    FactoryBot.create(:mentorship_period, mentor: mentor_period, mentee: ect_period, started_on: 1.year.ago, finished_on: nil)
   end
 
   describe '#fetch_etcs_and_mentors' do
@@ -18,9 +18,9 @@ describe School::ECTService do
 
       expect(result).to eq([
         {
-          ect: ect.full_name,
-          mentor: mentor.full_name,
-          status: School::ECTService::IN_PROGRESS,
+          ect: Teachers::Name.new(ect).full_name,
+          mentor: Teachers::Name.new(mentor).full_name,
+          status: Schools::Teacher::IN_PROGRESS,
         }
       ])
     end
